+++ src/utilities/tests/focus.test.ts
@@ -6,7 +6,8 @@ describe('focus', () => {
     it('calls blur on the currentTarget', () => {
       const currentTarget = document.createElement('button');
       currentTarget.blur = jest.fn();
-      handleMouseUpByBlurring({currentTarget} as MouseEvent<HTMLButtonElement>);
+      const mouseEvent = {currentTarget};
+      handleMouseUpByBlurring(mouseEvent as MouseEvent<HTMLButtonElement>);
       expect(currentTarget.blur).toHaveBeenCalled();
     });
   });