+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 * Consistently use contractions in public README (thanks [stefanmiodrag](https://github.com/stefanmiodrag) for the [original pull request](https://github.com/Shopify/polaris/pull/191)) ([#682](https://github.com/Shopify/polaris-react/pull/682))
 
 ### Enhancements
-* Added an option to show or hide unpublished products from the resource picker ([#628](https://github.com/Shopify/polaris-react/pull/628)
+* Added an option to show or hide unpublished products from the resource picker ([#628](https://github.com/Shopify/polaris-react/pull/628))
+* Minor updates to the Spinner documentation ([#696](https://github.com/Shopify/polaris-react/pull/696))
 
 ### Bug fixes
 * Fixed Layout component example description
+* Fixed SkeletonPage header appearing in embedded apps (thanks [rkbhochalya](https://github.com/rkbhochalya) for the [original issue](https://github.com/Shopify/polaris/issues/202))) ([#714](https://github.com/Shopify/polaris-react/pull/714))
 
 ## 1.7.0 - 2017-10-06
 ### Enhancements