+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Added an option to show or hide unpublished products from the resource picker ([#628](https://github.com/Shopify/polaris-react/pull/628))
 * Minor updates to the Spinner documentation ([#696](https://github.com/Shopify/polaris-react/pull/696))
 * Changed Popover component to use `react-transition-group` instead of our deprecated custom version in `@shopify/react-utilities` ([#718](https://github.com/Shopify/polaris-react/pull/718))
+* Extended support for disabled RadioButtons and Checkboxes rendered by ChoiceList ([#726](https://github.com/Shopify/polaris-react/pull/726))
 
 ### Bug fixes
 * Fixed Layout component example description